refactor get/set methods
